# RestaurantReservation.Backend/RestaurantReservation.Api/bin/Debug/netcoreapp3.1/Swaggerv1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Restaurant Reservation API v1",
    "description": "Restaurant Reservation API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Promotion/{id}": {
      "get": {
        "tags": [
          "Promotion"
        ],
        "summary": "Retrieve a specific restaurant Promotion information based on the Promotion ID given.",
        "operationId": "Get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Promotion ID",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Promotion ID",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A restaurant Promotion object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestaurantPromotionVM"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Error Retrieving a restaurant Promotion info from database",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "put": {
        "tags": [
          "Promotion"
        ],
        "summary": "Update an existing restaurant Promotion info in the database. This checks for duplicate name also.",
        "operationId": "Update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Promotion ID",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Promotion ID",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Updated Promotion Info",
          "content": {
            "application/json-patch+json; v=1.0": {
              "schema": {
                "$ref": "#/components/schemas/Promotion"
              }
            },
            "application/json; v=1.0": {
              "schema": {
                "$ref": "#/components/schemas/Promotion"
              }
            },
            "text/json; v=1.0": {
              "schema": {
                "$ref": "#/components/schemas/Promotion"
              }
            },
            "application/*+json; v=1.0": {
              "schema": {
                "$ref": "#/components/schemas/Promotion"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated Promotion object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Promotion"
                }
              }
            }
          },
          "400": {
            "description": "Invalid Promotion ID or Promotion Name exists message",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "Category Version Conflict message",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Error Updating a Promotion object information from database",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/Promotion/Check/{id}": {
      "get": {
        "tags": [
          "Promotion"
        ],
        "summary": "Retrieve a specific restaurant Promotion information based on the Promotion ID given.",
        "operationId": "CheckExistingByID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Promotion ID",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Promotion ID",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A restaurant Promotion object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestaurantPromotionVM"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Error Retrieving a restaurant Promotion info from database",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/Promotion": {
      "post": {
        "tags": [
          "Promotion"
        ],
        "summary": "Add a new promotion for a restaurant to the database. This checks for duplicate promotions name.",
        "operationId": "Add",
        "requestBody": {
          "description": "Promotion object",
          "content": {
            "application/json-patch+json; v=1.0": {
              "schema": {
                "$ref": "#/components/schemas/Promotion"
              }
            },
            "application/json; v=1.0": {
              "schema": {
                "$ref": "#/components/schemas/Promotion"
              }
            },
            "text/json; v=1.0": {
              "schema": {
                "$ref": "#/components/schemas/Promotion"
              }
            },
            "application/*+json; v=1.0": {
              "schema": {
                "$ref": "#/components/schemas/Promotion"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "A Boolean response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "500": {
            "description": "Error Adding a promotions object information into database",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/Promotion/{userId}/{id}": {
      "delete": {
        "tags": [
          "Promotion"
        ],
        "summary": "Remove an existing promotion information from database.",
        "operationId": "Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Promotion ID",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Promotion ID",
              "format": "int64"
            }
          },
          {
            "name": "userID",
            "in": "path",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "string",
              "description": "User ID",
              "nullable": true
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleting Promotion Result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Promotion record not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Error Deleting a Promotion object information from database",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/Promotion/PushEmail/{id}": {
      "get": {
        "tags": [
          "Promotion"
        ],
        "summary": "Push Email to the Member Subscriber based on Promotion ID given.",
        "operationId": "PushEmail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Promotion ID",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Promotion ID",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A Boolean response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "404": {
            "description": "Promotion record not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Restaurant": {
        "required": [
          "address",
          "isAvailable",
          "name"
        ],
        "type": "object",
        "properties": {
          "restaurantID": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "maxLength": 50,
            "type": "string"
          },
          "operatingHour": {
            "type": "string",
            "nullable": true
          },
          "workingDay": {
            "maxLength": 20,
            "type": "string",
            "nullable": true
          },
          "selectedWorkingDay": {
            "type": "integer",
            "format": "int32"
          },
          "startWorkingTime": {
            "type": "string",
            "format": "date-time"
          },
          "endWorkingTime": {
            "type": "string",
            "format": "date-time"
          },
          "imageUrl": {
            "maxLength": 500,
            "type": "string",
            "nullable": true
          },
          "address": {
            "maxLength": 100,
            "type": "string"
          },
          "isAvailable": {
            "type": "boolean"
          },
          "createdBy": {
            "maxLength": 80,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "editedBy": {
            "maxLength": 80,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "editedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "rowVersion": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "versionNo": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Promotion": {
        "required": [
          "content",
          "description",
          "endDate",
          "isAvailable",
          "name",
          "restaurantID",
          "startDate"
        ],
        "type": "object",
        "properties": {
          "promotionID": {
            "type": "integer",
            "format": "int64"
          },
          "restaurantID": {
            "type": "integer",
            "format": "int64"
          },
          "restaurant": {
            "$ref": "#/components/schemas/Restaurant"
          },
          "name": {
            "maxLength": 100,
            "type": "string"
          },
          "description": {
            "maxLength": 200,
            "type": "string"
          },
          "content": {
            "maxLength": 300,
            "type": "string"
          },
          "type": {
            "maxLength": 10,
            "type": "string",
            "nullable": true
          },
          "isAvailable": {
            "type": "boolean"
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "maxLength": 20,
            "type": "string",
            "nullable": true
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "editedBy": {
            "maxLength": 20,
            "type": "string",
            "nullable": true
          },
          "editedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "rowVersion": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "versionNo": {
            "type": "string",
            "nullable": true
          },
          "isEmailCreatedSent": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "RestaurantPromotionVM": {
        "type": "object",
        "properties": {
          "restaurantID": {
            "type": "integer",
            "format": "int64"
          },
          "restaurantName": {
            "type": "string",
            "nullable": true
          },
          "restaurantWorkingDay": {
            "type": "string",
            "nullable": true
          },
          "startWorkingTime": {
            "type": "string",
            "format": "date-time"
          },
          "endWorkingTime": {
            "type": "string",
            "format": "date-time"
          },
          "promotion": {
            "$ref": "#/components/schemas/Promotion"
          },
          "selectedType": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme. \r\n\r\n Enter 'Bearer' [space] and then your token in the text input below.\r\n\r\nExample: Bearer 12345abcdef",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}